backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
litellm==1.52.8
python-dotenv==1.0.1
pydantic==2.10.3
supabase==2.10.0
langchain==0.3.7
langchain-community==0.3.7
beautifulsoup4==4.12.3
requests==2.32.3
livekit-agents
livekit-plugins-openai
livekit-plugins-deepgram  
livekit-plugins-silero
langgraph
